+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp/Practise/Alex/Ilikefridges2.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.TeleOp;
+package org.firstinspires.ftc.teamcode.TeleOp.Practise.Alex;
 
 import com.qualcomm.hardware.rev.RevTouchSensor;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
