+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp/Practise/Felicity/sensortest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.TeleOp;
+package org.firstinspires.ftc.teamcode.TeleOp.Practise.Felicity;
 
 import com.qualcomm.hardware.rev.RevColorSensorV3;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
